Add jinja2 basic frontend
