+++ Dockerfile
@@ -6,6 +6,6 @@ RUN pip install pipenv
 WORKDIR /code
 COPY Pipfile Pipfile.lock /code/
 RUN pipenv install --system --deploy
-COPY backend /code/backend
+COPY app /code/backend
 
 CMD ["uvicorn", "backend.main:app", "--host", "0.0.0.0", "--port", "8000"]
