+++ app/db/models/task_logs.py
@@ -1,14 +1,14 @@
 from sqlalchemy import Column, String, Integer, func
 from sqlalchemy.orm import relationship
 
-from backend.db.base import Base
+from app.db.base import Base
 
 
 class TaskLogs(Base):
     __tablename__ = "task_logs"
 
     tid = Column(String, nullable=False, primary_key=True)
-    log_timestamp = Column(String, nullable=False, primary_key=True, default=func.now())
+    log_timestamp = Column(String, nullable=False, primary_key=True, default=func.now(tz="America/New_York"))
     level = Column(Integer, nullable=False, default=3)
     message = Column(String, nullable=False)
     task = relationship("Task")
