+++ app/main.py
@@ -1,0 +1,12 @@
+from fastapi import FastAPI
+from fastapi.staticfiles import StaticFiles
+
+from app.apis.base import api_router
+from app.core.config import settings
+from app.db.base import Base
+from app.db.session import engine
+
+app = FastAPI(title=settings.PROJECT_NAME, version=settings.PROJECT_VERSION)
+app.include_router(api_router)
+app.mount("/static", StaticFiles(directory="app/static"), name="static")
+Base.metadata.create_all(bind=engine)
