+++ app/static/css/not-so-stylish.css
@@ -1,0 +1,1 @@
+@import url("https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css");
