+++ app/templates/index.html
@@ -1,0 +1,16 @@
+{% extends "layout.html" %}
+
+{% block title %}
+    Home
+{% endblock %}
+
+{% block content %}
+    <div class="container">
+        <div class="row">
+            <div class="col-sm-12">
+                <p>Yeet</p>
+                <i class="bi-"></i>
+            </div>
+        </div>
+    </div>
+{% endblock %}
